+++ packages/components/atoms/f-form-field/test-utils/component-objects/f-form-field--visual.component.js
@@ -1,0 +1,13 @@
+const Page = require('@justeat/f-wdio-utils/src/page.object');
+
+module.exports = class FormField extends Page {
+    constructor () {
+        super('atom-folder', 'f-form-field--form-field-component');
+    }
+
+    get component () { return $('[data-test-id="formfield-container"]'); }
+
+    load () {
+        super.load(this.component);
+    }
+};