+++ packages/components/atoms/f-form-field/CHANGELOG.md
@@ -3,6 +3,15 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v4.6.1
+------------------------------
+*February 3, 2022*
+
+### Added
+- `FormField.stories.js` with all variants of f-form-field on a single storybook page.
+- `f-form-field--visual.component.js` and visual tests.
+
+
 v4.6.0
 ------------------------------
 *January 28, 2022*